chore(cli): Refactor files in prep for testing
