+++ cmd/list.go
@@ -17,10 +17,6 @@ limitations under the License.
 */
 
 import (
-	"context"
-
-	"github.com/meroxa/cli/display"
-
 	"github.com/spf13/cobra"
 )
 
@@ -32,182 +28,6 @@ var listCmd = &cobra.Command{
  resources, connectors, etc... You may also filter by type.`,
 }
 
-var listEndpointsCmd = &cobra.Command{
-	Use:     "endpoint",
-	Aliases: []string{"endpoints"},
-	Short:   "List endpoints",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), clientTimeOut)
-		defer cancel()
-
-		ends, err := c.ListEndpoints(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(ends)
-		} else {
-			display.PrintEndpointsTable(ends)
-		}
-
-		return nil
-	},
-}
-
-var listResourcesCmd = &cobra.Command{
-	Use:     "resources",
-	Short:   "List resources",
-	Aliases: []string{"resource"},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
-		defer cancel()
-
-		rr, err := c.ListResources(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(rr)
-		} else {
-			display.PrintResourcesTable(rr)
-		}
-		return nil
-	},
-}
-
-var listConnectorsCmd = &cobra.Command{
-	Use:     "connectors",
-	Short:   "List connectors",
-	Aliases: []string{"connector"},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
-		defer cancel()
-
-		connectors, err := c.ListConnectors(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(connectors)
-		} else {
-			display.PrintConnectorsTable(connectors)
-		}
-		return nil
-	},
-}
-
-var listResourceTypesCmd = &cobra.Command{
-	Use:     "resource-types",
-	Short:   "List resources-types",
-	Aliases: []string{"resource-type"},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
-		defer cancel()
-
-		resTypes, err := c.ListResourceTypes(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(resTypes)
-		} else {
-			display.PrintResourceTypesTable(resTypes)
-		}
-		return nil
-	},
-}
-
-var listPipelinesCmd = &cobra.Command{
-	Use:     "pipelines",
-	Short:   "List pipelines",
-	Aliases: []string{"pipeline"},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
-		defer cancel()
-
-		rr, err := c.ListPipelines(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(rr)
-		} else {
-			display.PrintPipelinesTable(rr)
-		}
-		return nil
-	},
-}
-
-var listTransformsCmd = &cobra.Command{
-	Use:     "transforms",
-	Short:   "List transforms",
-	Aliases: []string{"transform"},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		c, err := client()
-		if err != nil {
-			return err
-		}
-
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
-		defer cancel()
-
-		rr, err := c.ListTransforms(ctx)
-		if err != nil {
-			return err
-		}
-
-		if flagRootOutputJSON {
-			display.JSONPrint(rr)
-		} else {
-			display.PrintTransformsTable(rr)
-		}
-		return nil
-	},
-}
-
 func init() {
 	RootCmd.AddCommand(listCmd)
-
-	// Subcommands
-	listCmd.AddCommand(listResourcesCmd)
-	listCmd.AddCommand(listConnectorsCmd)
-	listCmd.AddCommand(listResourceTypesCmd)
-	listCmd.AddCommand(listPipelinesCmd)
-	listCmd.AddCommand(listTransformsCmd)
-	listCmd.AddCommand(listEndpointsCmd)
 }
