+++ cmd/list_pipelines.go
@@ -16,16 +16,41 @@ limitations under the License.
 package cmd
 
 import (
+	"context"
+
+	"github.com/meroxa/cli/display"
+
 	"github.com/spf13/cobra"
 )
 
-var resName, resType string
+var listPipelinesCmd = &cobra.Command{
+	Use:     "pipelines",
+	Short:   "List pipelines",
+	Aliases: []string{"pipeline"},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		c, err := client()
+		if err != nil {
+			return err
+		}
 
-var addCmd = &cobra.Command{
-	Use:   "add",
-	Short: "Add a resource to your Meroxa resource catalog",
+		ctx := context.Background()
+		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
+		defer cancel()
+
+		rr, err := c.ListPipelines(ctx)
+		if err != nil {
+			return err
+		}
+
+		if flagRootOutputJSON {
+			display.JSONPrint(rr)
+		} else {
+			display.PrintPipelinesTable(rr)
+		}
+		return nil
+	},
 }
 
 func init() {
-	RootCmd.AddCommand(addCmd)
+	listCmd.AddCommand(listPipelinesCmd)
 }
