+++ cmd/list_resources.go
@@ -13,19 +13,45 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package cmd
 
 import (
+	"context"
+
+	"github.com/meroxa/cli/display"
+
 	"github.com/spf13/cobra"
 )
 
-var resName, resType string
+var listResourcesCmd = &cobra.Command{
+	Use:     "resources",
+	Short:   "List resources",
+	Aliases: []string{"resource"},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		c, err := client()
+		if err != nil {
+			return err
+		}
 
-var addCmd = &cobra.Command{
-	Use:   "add",
-	Short: "Add a resource to your Meroxa resource catalog",
+		ctx := context.Background()
+		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
+		defer cancel()
+
+		rr, err := c.ListResources(ctx)
+		if err != nil {
+			return err
+		}
+
+		if flagRootOutputJSON {
+			display.JSONPrint(rr)
+		} else {
+			display.PrintResourcesTable(rr)
+		}
+		return nil
+	},
 }
 
 func init() {
-	RootCmd.AddCommand(addCmd)
+	listCmd.AddCommand(listResourcesCmd)
 }
