+++ cmd/list_transforms.go
@@ -13,19 +13,45 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package cmd
 
 import (
+	"context"
+
+	"github.com/meroxa/cli/display"
+
 	"github.com/spf13/cobra"
 )
 
-var resName, resType string
+var listTransformsCmd = &cobra.Command{
+	Use:     "transforms",
+	Short:   "List transforms",
+	Aliases: []string{"transform"},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		c, err := client()
+		if err != nil {
+			return err
+		}
 
-var addCmd = &cobra.Command{
-	Use:   "add",
-	Short: "Add a resource to your Meroxa resource catalog",
+		ctx := context.Background()
+		ctx, cancel := context.WithTimeout(ctx, clientTimeOut)
+		defer cancel()
+
+		rr, err := c.ListTransforms(ctx)
+		if err != nil {
+			return err
+		}
+
+		if flagRootOutputJSON {
+			display.JSONPrint(rr)
+		} else {
+			display.PrintTransformsTable(rr)
+		}
+		return nil
+	},
 }
 
 func init() {
-	RootCmd.AddCommand(addCmd)
+	listCmd.AddCommand(listTransformsCmd)
 }
