+++ docs/commands/meroxa.md
@@ -34,9 +34,7 @@ meroxa list resource-types
 * [meroxa logout](meroxa_logout.md)	 - logout of the Meroxa platform
 * [meroxa logs](meroxa_logs.md)	 - Print logs for a component
 * [meroxa open](meroxa_open.md)	 - Open in a web browser
-* [meroxa remove](meroxa_remove.md)	 - Remove a component
-* [meroxa update](meroxa_update.md)	 - Update a component
 * [meroxa update](meroxa_update.md)	 - Update a component
 * [meroxa version](meroxa_version.md)	 - Display the Meroxa CLI version
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
