+++ docs/commands/meroxa_add.md
@@ -20,4 +20,4 @@ Add a resource to your Meroxa resource catalog
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 * [meroxa add resource](meroxa_add_resource.md)	 - Add a resource to your Meroxa resource catalog
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
