+++ docs/commands/meroxa_add_resource.md
@@ -42,4 +42,4 @@ meroxa add resource slack --type url -u $WEBHOOK_URL
 
 * [meroxa add](meroxa_add.md)	 - Add a resource to your Meroxa resource catalog
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
