+++ docs/commands/meroxa_completion.md
@@ -57,4 +57,4 @@ meroxa completion [bash|zsh|fish|powershell]
 
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
