+++ docs/commands/meroxa_connect.md
@@ -42,4 +42,4 @@ meroxa connect --from <resource-name> --to <resource-name> [flags]
 
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
