+++ docs/commands/meroxa_create.md
@@ -27,4 +27,4 @@ including connectors.
 * [meroxa create endpoint](meroxa_create_endpoint.md)	 - Create an endpoint
 * [meroxa create pipeline](meroxa_create_pipeline.md)	 - Create a pipeline
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
