+++ docs/commands/meroxa_create_connector.md
@@ -38,4 +38,4 @@ meroxa create connector [<custom-connector-name>] --to pg2redshift --input order
 
 * [meroxa create](meroxa_create.md)	 - Create Meroxa pipeline components
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
