+++ docs/commands/meroxa_create_endpoint.md
@@ -36,4 +36,4 @@ meroxa create endpoint my-endpoint --protocol http --stream my-stream
 
 * [meroxa create](meroxa_create.md)	 - Create Meroxa pipeline components
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
