+++ docs/commands/meroxa_create_pipeline.md
@@ -24,4 +24,4 @@ meroxa create pipeline <name> [flags]
 
 * [meroxa create](meroxa_create.md)	 - Create Meroxa pipeline components
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
