+++ docs/commands/meroxa_describe.md
@@ -26,4 +26,4 @@ Describe a component of the Meroxa data platform, including resources and connec
 * [meroxa describe endpoint](meroxa_describe_endpoint.md)	 - Describe Endpoint
 * [meroxa describe resource](meroxa_describe_resource.md)	 - Describe resource
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
