+++ docs/commands/meroxa_describe_connector.md
@@ -23,4 +23,4 @@ meroxa describe connector [name] [flags]
 
 * [meroxa describe](meroxa_describe.md)	 - Describe a component
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
