+++ docs/commands/meroxa_list.md
@@ -30,4 +30,4 @@ List the components of the Meroxa platform, including pipelines,
 * [meroxa list resources](meroxa_list_resources.md)	 - List resources
 * [meroxa list transforms](meroxa_list_transforms.md)	 - List transforms
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
