+++ docs/commands/meroxa_list_resource-types.md
@@ -23,4 +23,4 @@ meroxa list resource-types [flags]
 
 * [meroxa list](meroxa_list.md)	 - List components
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
