+++ docs/commands/meroxa_logs.md
@@ -20,4 +20,4 @@ Print logs for a component
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 * [meroxa logs connector](meroxa_logs_connector.md)	 - Print logs for a connector
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
