+++ docs/commands/meroxa_logs_connector.md
@@ -23,4 +23,4 @@ meroxa logs connector <name> [flags]
 
 * [meroxa logs](meroxa_logs.md)	 - Print logs for a component
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
