+++ docs/commands/meroxa_open.md
@@ -20,4 +20,4 @@ Open in a web browser
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 * [meroxa open billing](meroxa_open_billing.md)	 - Open your billing page in a web browser
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
