+++ docs/commands/meroxa_open_billing.md
@@ -23,4 +23,4 @@ meroxa open billing [flags]
 
 * [meroxa open](meroxa_open.md)	 - Open in a web browser
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
