+++ docs/commands/meroxa_remove.md
@@ -28,4 +28,4 @@ Deprovision a component of the Meroxa platform, including pipelines,
 * [meroxa remove pipeline](meroxa_remove_pipeline.md)	 - Remove pipeline
 * [meroxa remove resource](meroxa_remove_resource.md)	 - Remove resource
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
