+++ docs/commands/meroxa_remove_endpoint.md
@@ -23,4 +23,4 @@ meroxa remove endpoint <name> [flags]
 
 * [meroxa remove](meroxa_remove.md)	 - Remove a component
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
