+++ docs/commands/meroxa_update.md
@@ -26,4 +26,4 @@ Update a component of the Meroxa platform, including connectors
 * [meroxa update pipeline](meroxa_update_pipeline.md)	 - Update pipeline state
 * [meroxa update resource](meroxa_update_resource.md)	 - Update a resource
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
