+++ docs/commands/meroxa_update_pipeline.md
@@ -24,4 +24,4 @@ meroxa update pipeline <name> --state <pause|resume|restart> [flags]
 
 * [meroxa update](meroxa_update.md)	 - Update a component
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
