+++ docs/commands/meroxa_update_resource.md
@@ -30,4 +30,4 @@ meroxa update resource <resource-name> [flags]
 
 * [meroxa update](meroxa_update.md)	 - Update a component
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
