+++ docs/commands/meroxa_version.md
@@ -23,4 +23,4 @@ meroxa version [flags]
 
 * [meroxa](meroxa.md)	 - The Meroxa CLI
 
-###### Auto generated by spf13/cobra on 12-Mar-2021
+###### Auto generated by spf13/cobra on 13-Mar-2021
